feat: restructure into two components
